+++ sdk/storage/azure-storage-blob/src/main/java/com/azure/storage/blob/specialized/BlobClientBase.java
@@ -1624,7 +1624,9 @@ public class BlobClientBase {
                 finalRequestConditions.getLeaseId(), deleteBlobSnapshotOptions,
                 finalRequestConditions.getIfModifiedSince(), finalRequestConditions.getIfUnmodifiedSince(),
                 finalRequestConditions.getIfMatch(), finalRequestConditions.getIfNoneMatch(),
-                finalRequestConditions.getTagsConditions(), null, null, finalContext);
+                finalRequestConditions.getTagsConditions(), null, null,
+                finalRequestConditions.getAccessTierIfModifiedSince(),
+                finalRequestConditions.getAccessTierIfUnmodifiedSince(), finalContext);
 
         return sendRequest(operation, timeout, BlobStorageException.class);
     }